Feat: add tests and implement logic contain
This commit refactors 'Criteria' domain class by adding new methods and updating existing ones. Changes also include modifications in related test classes to match these updates. The main goal of this refactor is to improve the overall functionalities of the 'Criteria' class, and thereby, increase the maintainability and readability of the code.